week10/requirements.txt
matplotlib==3.7.4
numpy==1.24.4
opencv_python==4.9.0.80
Pillow==10.3.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
